# public/openapi.yaml
openapi: 3.0.3
info:
  title: Anikoto Scraper API
  description: |
    A high-performance REST API for scraping anime data from **anikoto.net** — built with Next.js, Cheerio, and in-memory caching.

    ## Response Format
    All endpoints return a consistent JSON envelope:
    ```json
    { "ok": true, "data": { ... }, "cached": true }
    ```
    On error:
    ```json
    { "ok": false, "message": "Error description" }
    ```

    ## Caching
    | Endpoint | TTL |
    |----------|-----|
    | `/api/home` | 5 minutes |
    | `/api/anime/:slug` | 30 minutes |
    | `/api/search` | 2 minutes |
    | `/api/filter` | 5 minutes |
    | `/api/schedule` | 1 hour |
    | Episodes | 10 minutes |

    Add `?refresh=1` to any cached endpoint to bypass the cache.
  version: "1.0.0"
  contact:
    name: Teramoto
servers:
  - url: https://anikoto-scrap-api.vercel.app/api
    description: Production server
  - url: /api
    description: Local development server
...
paths:
...
  /search:
    get:
      tags: [Browse]
      summary: Search anime by keyword
      operationId: searchAnime
      parameters:
        - name: keyword
          in: query
          required: true
          description: Search keyword
          schema:
            type: string
            example: one piece
        - $ref: '#/components/parameters/refresh'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  cached:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/SearchResult'
              example:
                ok: true
                cached: false
                data:
                  keyword: one piece
                  totalResults: 42
                  results: []
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/Error'
...
  /status:
    get:
      tags: [Browse]
      summary: Browse anime by airing status
      operationId: getByStatus
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [currently-airing, finished-airing, not-yet-aired]
            default: currently-airing
          description: Airing status
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
      responses:
        '200':
          description: Status listing
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  cached:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnimeCard'
        '500':
          $ref: '#/components/responses/Error'
...
  /type/{type}:
    get:
      tags: [Browse]
      summary: Browse anime by media type
      operationId: getByType
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [tv, movie, ova, ona, special, music]
          description: Media type
          example: movie
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
      responses:
        '200':
          description: Type listing
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  cached:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnimeCard'
        '500':
          $ref: '#/components/responses/Error'
...
  /proxy:
    get:
      tags: [Watch]
      summary: Streaming proxy
      description: >
        Internal proxy to bypass Cloudflare and CORS restrictions for m3u8 video streams and subtitle files.
        Rewrites segment URLs within m3u8 playlists to keep them routed through the proxy.
      operationId: proxyStream
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: The target m3u8 or subtitle URL to proxy
          example: "https://cdn.mewstream.buzz/.../master.m3u8"
        - name: referer
          in: query
          required: false
          schema:
            type: string
            format: uri
          description: Referer header to bypass hotlink protection
          example: "https://megaplay.buzz/"
      responses:
        '200':
          description: Proxied stream data
          content:
            application/vnd.apple.mpegurl:
              schema:
                type: string
                description: m3u8 playlist with rewritten segment URLs
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/Error'
components:
  parameters:
...
    refresh:
      name: refresh
      in: query
      required: false
      schema:
        type: string
        enum: ["1"]
      description: Set to `1` to bypass cache and force a fresh scrape
  schemas:
...
    EpisodeStatus:
      type: object
      properties:
        sub:
          type: integer
          nullable: true
        dub:
          type: integer
          nullable: true
        total:
          type: integer
          nullable: true
    AnimeCard:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        titleJp:
          type: string
        image:
          type: string
          format: uri
        href:
          type: string
        type:
          type: string
        episodes:
          $ref: '#/components/schemas/EpisodeStatus'
        date:
          type: string
        score:
          type: number
        totalEpisodes:
          type: integer

    # ─── Home ──────────────────────────────────────────────────────
...
    SearchResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/AnimeCard'
        keyword:
          type: string
        totalResults:
          type: integer
...
    ApiError:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
          example: false
        message:
          type: string
          example: "Internal server error"
  responses:
    Error:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            ok: false
            message: Failed to fetch data from source
    BadRequest:
      description: Bad request — missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            ok: false
            message: Missing required parameter
...